Add window directional commands
This adds the directional commands for window control, useful for
hotkeys and eliminating the need for additional software. Per the
documentation[1] the focus commands will change the focused window
in the direction specified, the swap commands exchange size and
position, and the warp commands re-insert the current actively focused
window in the direction specified.